requirements.txt
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
typer>=0.12
streamlit>=1.28
